+++ src/main/java/com/isco/Blog/Controller/BlogGroupController.java
@@ -1,9 +1,6 @@
 package com.isco.Blog.Controller;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,32 +14,46 @@ import com.isco.Blog.Service.BlogGroupService;
 public class BlogGroupController {
 
 	@Autowired
-    BlogGroupService blogGroupService;
-	
-	//新建博客组（用户）
-	@RequestMapping(path="/addBlogGroup",method=RequestMethod.POST)
+	BlogGroupService blogGroupService;
+
+	// 新建博客组（用户）
+	@RequestMapping(path = "/addBlogGroup", method = RequestMethod.POST)
 	public int addBlogGroup(@RequestBody BlogGroup bGroup) {
-		if(bGroup.getName()==null)
-			return -1;
+		if (bGroup.getName() == null || bGroup.getUserId() == null)
+			return -2;
 		return blogGroupService.insertGroup(bGroup);
 	}
-	//更改博客组
-	@RequestMapping(path="/updateBlogGroup",method=RequestMethod.POST)
+
+	// 更改博客组
+	@RequestMapping(path = "/updateBlogGroup", method = RequestMethod.POST)
 	public int updateBlogType(@RequestBody BlogGroup bGroup) {
-		if(bGroup.getId()==null || bGroup.getName()==null)
-			return -1;
+		if (bGroup.getName() == null)
+			return -2;
 		return blogGroupService.updateGroupByPrimaryKey(bGroup);
 	}
-	//删除博客组
-	@RequestMapping(path="/delBlogGroup",method=RequestMethod.DELETE)
-	public int delBlogGroup(int id) {
+
+	// 删除博客组
+	@RequestMapping(path = "/delBlogGroup", method = RequestMethod.DELETE)
+	public int delBlogGroup(Integer id) {
+		if(id==null)
+			return -1;
 		return blogGroupService.deleteGroupByPrimaryKey(id);
 	}
-	
-	//列出所有博客组(传输参数待定)
-	@RequestMapping(path="/getBlogGroup",method=RequestMethod.GET)
-	public List<BlogGroup> ListBlogType(int userId) {
+
+	// 列出所有博客组和博客
+	@RequestMapping(path = "/getBlogGroup", method = RequestMethod.GET)
+	public List<BlogGroup> ListBlogType(Integer userId) {
+		if(userId==null)
+				return null;
 		return blogGroupService.selectByUserId(userId);
+	}
+
+	// 列出所有博客组和博客
+	@RequestMapping(path = "/getBlogGroupJustList", method = RequestMethod.GET)
+	public List<BlogGroup> ListBlogTypeWithoutBlog(Integer userId) {
+		if(userId==null)
+			return null;
+		return blogGroupService.selectByUserIdJustList(userId);
 	}
 
 }
